Delete deprecated authenticator
Old Oauth2 implementation that's no longer in use by the base classes
